Cleanup and sort all the using
There are few reasons for removing unused using(s)/namespaces, besides
coding preference:

removing the unused using clauses in a project, can make the compilation
faster because the compiler has fewer namespaces to look-up types to
resolve. (this is especially true for C# 3.0 because of extension
methods, where the compiler must search all namespaces for extension
methods for possible better matches, generic type inference and lambda
expressions involving generic types)
can potentially help to avoid name collision in future builds when new
types are added to the unused namespaces that have the same name as some
types in the used namespaces.
will reduce the number of items in the editor auto completion list when
coding, posibly leading to faster typing (in C# 3.0 this can also reduce
the list of extension methods shown)